% Generated by roxygen2 (4.0.1): do not edit by hand
\name{carouselPanel}
\alias{carouselPanel}
\title{Layout a Carousel Panel}
\usage{
carouselPanel(..., auto.advance = FALSE)
}
\arguments{
\item{...}{Elements of the carousel.  Can be multiple elements wrapped in a
div(), or single output elements.}

\item{auto.advance}{Wheather the carousel should automatically advance to the
  next element, of it it should only advance when the user clicks the navigation
  arrows.}
}
\description{
Adds a carousel panel to the current shiny app.
}
\examples{
# Define UI
shinyUI(fluidPage(

  # Application title
  titlePanel("Hello Shiny!"),

  sidebarLayout(

    # Sidebar with a slider input
    sidebarPanel(
      sliderInput("obs",
                  "Number of observations:",
                  min = 0,
                  max = 1000,
                  value = 500)
    ),

    # Show a plot of the generated distribution
    mainPanel(
      carouselPanel(
         plotOutput("distPlot"),
         plotOutput("xyPlot"),
         auto.advance=TRUE
       )
    )
  )
))
}

