% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grkShinyThings-package.R
\docType{package}
\name{grkShinyThings-package}
\alias{grkShinyThings}
\alias{grkShinyThings-package}
\title{grkShinyThings: Reusable Shiny Modules and Other Shiny Things}
\description{
A collection of re-usable Shiny modules that can be
    dropped into Shiny apps. Also includes additional helper functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/GerkeLab/grkShinyThings}
  \item Report bugs at \url{https://github.com/GerkeLab/grkShinyThings/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
