% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localStorage.R
\name{localStorage}
\alias{localStorage}
\alias{localStorage_store}
\alias{localStorage_remove}
\alias{localStorage_clear}
\alias{use_localStorage}
\title{Access the Browser's Local Storage}
\usage{
localStorage_store(..., overwrite = FALSE, .list = NULL,
  session = shiny::getDefaultReactiveDomain())

localStorage_remove(key, session = shiny::getDefaultReactiveDomain())

localStorage_clear(clear = TRUE,
  session = shiny::getDefaultReactiveDomain())

use_localStorage()
}
\description{
Store and retrieve items from the browser's local storage.
}
\section{Functions}{
\itemize{
\item \code{localStorage_store}: Store key-value pairs in the browser's local storage.

\item \code{localStorage_remove}: Remove a key-value pair from the browser's local storage.

\item \code{localStorage_clear}: Clear the browser's local storage.

\item \code{use_localStorage}: Use the browser's local storage in your Shiny app.
Include this function in your UI to expose the input \code{"_localStorage"} in
your app and to enable the \code{localStorage_*()} functions.
}}

