% Generated by roxygen2 (4.0.1): do not edit by hand
\name{modalButton}
\alias{modalButton}
\title{Modal button}
\usage{
modalButton(label, styleclass = "", size = "", block = F, icon = NULL,
  icon.library = c("font awesome", "bootstrap"), css.class = "", modal.id,
  ...)
}
\arguments{
\item{label}{The contents of the button--usually a text label, but you could
also use any other HTML, like an image.}

\item{styleclass}{The Bootstrap styling class of the button--options are
primary, info, success, warning, danger, inverse, link or blank}

\item{size}{The size of the button--options are large, small, mini}

\item{block}{Whehter the button should fill the block}

\item{icon}{Display an icon for the button}

\item{specify}{icon set to use
\url{http://www.fontawesome.io/icons} or
\url{http://getbootstrap.com/2.3.2/base-css.html#icons}}

\item{css.class}{Any additional CSS class one wishes to add to the action
button}

\item{modal.id}{The id of the modal that the button should toggle.}
}
\description{
Creates a button that launches a modal.
}
\seealso{
Other ShinyThings elements: \code{\link{numericInput}};
  \code{\link{passwordInput}};
  \code{\link{updateActionButton}};
  \code{\link{updatePasswordInput}}
}

