% Generated by roxygen2 (4.0.1): do not edit by hand
\name{modalPanel}
\alias{modalPanel}
\title{Layout a Modal}
\usage{
modalPanel(outputId, header = NULL, body = NULL, footer = NULL)
}
\arguments{
\item{outputId}{Element id.}

\item{header}{Text to use as the modal header.}

\item{body}{Elements to use as the modal body.}

\item{footer}{Elements to use as the modal footer.}
}
\description{
Adds a modal to the current shiny app.
}

