% Generated by roxygen2 (4.0.1): do not edit by hand
\name{numericInput}
\alias{numericInput}
\title{Numeric Input}
\usage{
numericInput(inputId, label, value, min = NA, max = NA, step = NA)
}
\arguments{
\item{inputId}{Specifies the input slot that will be used to access the
value.}

\item{label}{The contents of the button--usually a text label, but you could
also use any other HTML, like an image.}

\item{value}{of the input element}

\item{minimum}{of the input element}

\item{maximum}{of the input element}

\item{step}{between values of the input element}
}
\description{
Creates a numeric input element
}
\seealso{
Other ShinyThings elements: \code{\link{passwordInput}};
  \code{\link{updateActionButton}};
  \code{\link{updatePasswordInput}}
}

