% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pager.R
\name{pager}
\alias{pager}
\alias{paginationUI}
\alias{pagerUI}
\alias{pagerDemo}
\title{Pagination and Pager Module}
\usage{
pager(id, n_items, page_break = 20)

paginationUI(id, width = 6, ..., offset = 3)

pagerUI(id, label_prev = "Previous", label_next = "Next",
  centered = TRUE, class = NULL)

pagerDemo()
}
\arguments{
\item{id}{The shared \code{id} of the \code{pagerUI()} and the \code{pager()} module}

\item{n_items}{The number of items in the collection of items to be paged}

\item{page_break}{The maximum number of items in each page}

\item{width}{The width of the containing column}

\item{...}{Additional elements included in the containg column, added before
the pagniation buttons}

\item{offset}{The offest of the containing column}
}
\description{
A pager module to add pagination to any collection of items. Given a
number of items and a page break, the pagination module returns the
indexes of the items on the currently selected page.
}
\section{Functions}{
\itemize{
\item \code{paginationUI}: The user interface for the pager module. The pager is
contained in a single column in a \code{fluidRow()}.

\item \code{pagerUI}: Next/previous buttons that can be used in addition to,
or in place of \code{\link[=paginationUI]{paginationUI()}}.

\item \code{pagerDemo}: Example app demonstrating usage of the pager module.
}}

\references{
\url{https://getbootstrap.com/docs/3.3/components/#pagination}
}
