% Generated by roxygen2 (4.0.1): do not edit by hand
\name{passwordInput}
\alias{passwordInput}
\title{Password Input}
\usage{
passwordInput(inputId, label, value = "")
}
\arguments{
\item{inputId}{Specifies the input slot that will be used to access the
value.}

\item{label}{The contents of the button--usually a text label, but you could
also use any other HTML, like an image.}

\item{value}{of the input element}
}
\description{
Creates a password input element that obscures any text entered.  The value is
not encrypted before transmission to the server.  This is intended to provide
the look and feel of typical password field to those who know that there is
absolutely NO SECURITY in using it within your app.  The data transmitted is
PLAIN, UNENCRYPTED TEXT.
}
\seealso{
Other ShinyThings elements: \code{\link{numericInput}};
  \code{\link{updateActionButton}};
  \code{\link{updatePasswordInput}}
}

