% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyThings-package.R
\docType{package}
\name{shinyThings-package}
\alias{shinyThings}
\alias{shinyThings-package}
\title{shinyThings: Reusable Shiny Modules and Other Shiny Things}
\description{
A collection of re-usable Shiny modules that can be
    dropped into Shiny apps. Also includes additional helper functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gadenbuie/shinyThings}
  \item Report bugs at \url{https://github.com/gadenbuie/shinyThings/issues}
}

}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{garrick@adenbuie.com} (0000-0002-7111-0077)

}
\keyword{internal}
