% Generated by roxygen2 (4.0.2): do not edit by hand
\name{updatePasswordInput}
\alias{updatePasswordInput}
\title{Update Password Input}
\usage{
updatePasswordInput(session, inputId, label = NULL, value = NULL)
}
\arguments{
\item{session}{identifier}

\item{inputId}{Specifies the input slot that will be used to access the
value.}

\item{new}{label The contents of the button--usually a text label, but you could
also use any other HTML, like an image.}

\item{new}{value of the input element}
}
\description{
Updates a password input element.
}
\seealso{
Other ShinyThings elements: \code{\link{modalButton}};
  \code{\link{numericInput}}; \code{\link{passwordInput}};
  \code{\link{updateActionButton}}
}

