% Generated by roxygen2 (4.0.1): do not edit by hand
\name{actionButton}
\alias{actionButton}
\title{Action button/link}
\usage{
actionButton(inputId, label, type = c("", "primary", "info", "success",
  "warning", "danger", "inverse", "link"), size = c("", "large", "small",
  "mini"), block = F, icon = NULL, icon.library = c("font awesome",
  "bootstrap"), class = "", ...)
}
\arguments{
\item{inputId}{Specifies the input slot that will be used to access the
value.}

\item{label}{The contents of the button--usually a text label, but you could
also use any other HTML, like an image.}

\item{type}{The Bootstrap style of the button--options are
primary, info, success, warning, danger, inverse, link or blank}

\item{size}{The size of the button--options are large, small, mini}

\item{block}{Whether the button should fill the block}

\item{icon}{An optional \code{\link{icon}} to appear on the button.}

\item{icon.library}{specify icon set to use
\url{http://www.fontawesome.io/icons} or
\url{http://getbootstrap.com/2.3.2/base-css.html#icons}}

\item{class}{Any additional CSS class one wishes to add to the action
button}

\item{...}{Named attributes to be applied to the button or link.}
}
\description{
Creates an action button whose value is initially zero, and increments by one
each time it is pressed.
}
\examples{
\dontrun{
# In server.R
output$distPlot <- renderPlot({
  # Take a dependency on input$goButton
  input$goButton

  # Use isolate() to avoid dependency on input$obs
  dist <- isolate(rnorm(input$obs))
  hist(dist)
})

# In ui.R
actionButton("goButton", "Go!", icon="space-shuttle", type="success")
}
}
\seealso{
Other input elements: \code{\link{fileInput}}
}

