% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button-group.R
\name{buttonGroup}
\alias{buttonGroup}
\title{A Bootstrap Button Group Input}
\usage{
buttonGroup(inputId, options, btn_class = "btn-default",
  selected = NULL, multiple = FALSE, aria_label = NULL, ...)
}
\arguments{
\item{inputId}{The input id}

\item{options}{A named vector of options and labels. The name is the label
that will appear on the button and the value is the id of the input that is
returned from the Shiny modules.}

\item{btn_class}{A single class applied to each individual button, or a
vector of button classes for each button (must be same length as \code{options})}

\item{selected}{The buttons, by button value, that should be activated.}

\item{multiple}{By default, only a single button may be toggled at a time.
If \code{multiple} is \code{TRUE}, then \code{buttonGroup()} returns a character vector
of the selected button values.}

\item{...}{Passed to \code{\link[htmltools:div]{htmltools::div()}}}
}
\description{
This input operates like a \code{\link[shiny:radioButtons]{shiny::radioButtons()}} or
\code{\link[shiny:checkboxGroupInput]{shiny::checkboxGroupInput()}} input.
}
